requirements.txt
Django>=4.1.5,<5.0
djangorestframework>=3.14.0
psycopg2-binary
